test: split space and variable tests
